Proof of concept, remove questions
This is a working proof of concept to demonstrate how to avoid
some questions.
The goal ist the following:
Remove the question for grok, if it can be added or removed
with a single modification.
This goal has been reached, you only need to uncomment
the five.grok line in setup.py. There is conditional code in the
package zcml that will only load grok, if it is available.