Cleanup generated MANIFEST.in, add setup.cfg.
There is no need to include *cfg files, nothing is done with it.

The MANIFEST.in file is included automatically, so no need to spell it out.

We must globally exclude *.pyc.

In short, the check-manifest command should not complain.  This is a
stand-alone command and an optional helper package for zest.releaser,
called during prerelease.

Also, this fixes an error on Windows, noticed by check-manifest:
ERROR: Trailing slashes are not allowed in MANIFEST.in on Windows: docs/