Added a post_ask hook to set variables always.
This is called after all questions have been asked, also when you are
in non-interactive mode.

This way we do not have to fill the answers files with variables that
are calculated when in interactive mode.

Removed those calculated answers from these files.  This also means
that another problem is solved: 'plone.is_plone5 = False' in such a
test file was taken by mr.bob to be a string 'False', which in boolean
means: True...  So the tests were generating some files and options
that were only meant for Plone 5, also when we specified the plone
version as 4.3.6.