Fixed i18n attributes for View/Edit actions in dexterity type xml.
I got a warning when running the update script:

$ ./update.sh
Warning: msgid 'description' in ../profiles/default/types/Graph.xml already exists with a different default (bad: Edit, should be: View)
The references for the existent value are:
../profiles/default/types/Graph.xml

This pointed to the real problem. We were using
`i18n:attributes="title description"` which means:
translate the `title` attribute and give it message id `description`.
With the new code, we say: translate both the `title` and `description` attributes,
without giving an explicit message id.